chore: improve Popover exports
